+++ src/webview/components/PersonInfo.tsx
@@ -10,21 +10,40 @@ export interface PersonInfoProps {
     label: string;
 }
 
-export function getRelativeTimeString(timestamp: string): string {
+export function getRelativeTimeString(timestamp: string, now: number | null = null): string {
     const time = new Date(timestamp).getTime();
     if (isNaN(time)) return timestamp;
 
-    const now = Date.now();
-    const diff = now - time;
+    // If we assume a year is 365.25 days:
+    const SECONDS_PER_YEAR = 365.25 * 24 * 60 * 60;
+    const SECONDS_PER_MONTH = SECONDS_PER_YEAR / 12;
 
-    const seconds = Math.floor(diff / 1000);
+    const diffMs = (now ?? Date.now()) - time;
+    // This function only works for timestamps in the past.
+    // (We could make it work for timestamps in the future, but that isn't a valid input for now.)
+    if (diffMs < 0) return timestamp;
+
+    // Display a single unit of time from seconds to years.
+    const seconds = Math.floor(diffMs / 1000);
     const minutes = Math.floor(seconds / 60);
     const hours = Math.floor(minutes / 60);
     const days = Math.floor(hours / 24);
-
-    if (days > 0) return `${days} day${days > 1 ? 's' : ''} ago`;
-    if (hours > 0) return `${hours} hour${hours > 1 ? 's' : ''} ago`;
-    if (minutes > 0) return `${minutes} minute${minutes > 1 ? 's' : ''} ago`;
+    const weeks = Math.floor(days / 7);
+    const months = Math.floor(seconds / SECONDS_PER_MONTH);
+    const years = Math.floor(seconds / SECONDS_PER_YEAR);
+    for (const [x, unit] of [
+        [years, 'year'],
+        [months, 'month'],
+        [weeks, 'week'],
+        [days, 'day'],
+        [hours, 'hour'],
+        [minutes, 'minute'],
+        [seconds, 'second'],
+    ] satisfies Array<[number, string]>) {
+        if (x > 0) {
+            return `${x} ${unit}${x > 1 ? 's' : ''} ago`;
+        }
+    }
     return 'just now';
 }